refactor(cli/platform): optimize structure for CI runs
Closes https://github.com/prisma/team-orm/issues/894
