+++ packages/cli/src/platform/auth/show.ts
@@ -1,7 +1,7 @@
 import { arg, Command, formatTable, isError } from '@prisma/internals'
 import { green } from 'kleur/colors'
 
-import { getPlatformTokenOrThrow, platformParameters, platformRequestOrThrow } from '../../utils/platform'
+import { getPlatformTokenOrThrow, platformParameters, platformRequestOrThrow } from '../platformUtils'
 
 export class Show implements Command {
   public static new(): Show {
